refactor: move components to src directory and update import paths
