+++ src/app/api/auth/nonce/route.ts
