+++ src/app/workouts/details/page.tsx
