+++ src/app/workouts/page.tsx
@@ -2,8 +2,8 @@
 
 import React from "react";
 import { useRouter } from "next/navigation";
-import { workoutLists } from "../Mocks/WorkoutsList";
-import { WorkoutContainer } from "@/components/Containers/WorkoutContainer";
+import { workoutLists } from "@/src/constants/mocks/WorkoutsList";
+import { WorkoutContainer } from "@/src/components/Containers/WorkoutContainer";
 
 export default function WorkoutsPage() {
   const router = useRouter();
