+++ src/components/Cards/DetailedWorkoutCard.tsx
