+++ src/components/Pay/index.tsx
