+++ src/components/WalletAuth/index.ts
